hoertje123

+++ core/src/main/java/com/groeps33/valley/entity/Entity.java
@@ -15,8 +15,8 @@ public abstract class Entity {
     protected int maxHp;
     protected int defence;
     protected int attackDamage;
-    protected int attackSpeedInterval;
-    protected int moveSpeed;
+    protected float attackSpeedInterval;
+    protected float moveSpeed;
     protected Vector2 location;
     protected Vector2 velocity;
